pythonrequirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib[bcrypt]
pymongo
python-dotenv
python-multipart
Pillow
pydantic[email]
email-validator
bcrypt==3.2.2
pandas
openpyxl
xlrd
numpy
scipy
statsmodels
python-dateutil
